renamed mip_rd to mip
