+++ cg/cli/workflow/__init__.py
@@ -3,5 +3,5 @@
 from cg.cli.workflow.workflow import workflow
 from cg.cli.workflow.balsamic import balsamic
 from cg.cli.workflow.microsalt import microsalt
-from cg.cli.workflow.mip_rd_dna import mip_dna
-from cg.cli.workflow.mip_rd_rna import mip_rna
+from cg.cli.workflow.mip_dna import mip_dna
+from cg.cli.workflow.mip_rna import mip_rna
