+++ tests/cli/workflow/mip_dna/test_cli_mip_dna_start.py
@@ -1,6 +1,6 @@
 """This script tests the cli methods to create prerequisites and start a mip-dna analysis"""
 import logging
-from cg.cli.workflow.mip_rd_dna import start
+from cg.cli.workflow.mip_dna import start
 
 EXIT_SUCCESS = 0
